# Remove message boxing for performance improvement