This is an extension of the message boxing approach proposed in PR #370. This however continues the improvements to remove additional type information that isn't necessary.

This diff includes the removal of the BoxedMessage struct in favor of a crate-private RactorMessage trait which adds the necessary extension methods, only within the `ractor`
context so as the privatize that functionality, as it was an unintentional leak anyways. This of course is a semver break as a public type is going away, but it should have been unused
anyways in all contexts. Additionally `BoxedDowncastErr` is renamed to `DowncastErr` as we are no longer boxing the message, but again this error variant shouldn't really be used publicly.

Will re-run the performance benchmarks present in #370 and past as a PR comment once ready