@slawlor 2025-08-20
## Performance tests

`cargo clean && cargo bench -p ractor --profile release --bench actor`

### Main branch
**Excluding the output port tests** (because they take forever and should be isolated anyways, which this PR does properly)

```
Creation of 100 actors  time:   [227.75 µs 234.65 µs 240.96 µs]
Creation of 10000 actors
                        time:   [18.840 ms 19.632 ms 20.495 ms]
Waiting on 100 actors to process first message
                        time:   [865.42 µs 898.55 µs 937.11 µs]
Waiting on 1000 actors to process first message
                        time:   [10.389 ms 10.985 ms 11.657 ms]
Waiting on 100000 messages to be processed
                        time:   [13.157 ms 13.258 ms 13.372 ms]
  ````

### this branch

```
Creation of 100 actors  time:   [271.34 µs 282.47 µs 293.46 µs]
Creation of 10000 actors
                        time:   [27.574 ms 28.371 ms 29.157 ms]
Waiting on 100 actors to process first message
                        time:   [940.67 µs 974.26 µs 1.0130 ms]
Waiting on 1000 actors to process first message
                        time:   [9.6253 ms 9.8824 ms 10.179 ms]
Waiting on 100000 messages to be processed
                        time:   [12.740 ms 13.294 ms 14.012 ms]
  ```