@slawlor 2025-08-20
Added a new benchmark for message sending to assess if sending (but not receiving) messages is improved with this change.

## This branch

```
Sending of 1000 messages to an actor
                        time:   [46.212 µs 47.399 µs 48.730 µs]
Sending of 100000 messages to an actor
                        time:   [5.1381 ms 5.3625 ms 5.6080 ms]
  ```
  
 ## main

 Patching this benchmark suite on main and running there yields: 
 
 ```
 Sending of 1000 messages to an actor
                        time:   [46.876 µs 47.673 µs 48.605 µs]
                        change: [-0.8271% +2.5668% +6.0038%] (p = 0.15 > 0.05)
Sending of 100000 messages to an actor
                        time:   [6.3139 ms 6.5667 ms 6.8343 ms]
                        change: [+15.550% +22.456% +29.592%] (p = 0.00 < 0.05)
 ```